Struct raw::compounds::Subclip

pub struct Subclip { /* fields omitted */ }

A struct used to represent a subset of audio data within a larger set via reference. Good for working with smaller chunks of data out of a SampleArray.

Methods

impl Subclip

pub fn new<S: Time, T: Time>(
clip: Arc<Clip>,
start_time: S,
duration_time: T
) -> Option<Arc<Self>>

The split time (start_time + duration_time) must be within the duration of the clip, else None is returned.

pub fn from_start<T: Time>(
clip: Arc<Clip>,
duration_time: T
) -> Option<Arc<Self>>

Returns a SubClip from the beginning of the source clip to a specified time into the clip.

pub fn to_end<T: Time>(clip: Arc<Clip>, start_time: T) -> Option<Arc<Self>>

Returns a SubClip from the specified start time to the end of the source clip.

pub fn split<T: Time>(
clip: Arc<Clip>,
split_time: T
) -> Option<(Arc<Self>, Arc<Self>)>

Returns two SubClips, one from the beginning of the source clip to the split point, and one from the split point to the end of the source clip.