Trait raw::outline::Time

pub trait Time: Sized {
    fn to_samples(&self, sample_rate: u32) -> u64;

    fn from_samples(&self, num_samples: u64, sample_rate: u32) -> Self;
}

moments and durations are represented by types that implement Time. Convertible to/from the equivalent number of samples. This allows time to be represented in different ways, e.g. measures+beats, seconds+millis, etc.

Required Methods

fn to_samples(&self, sample_rate: u32) -> u64

the number of samples that takes up the same amount of time as self

fn from_samples(&self, num_samples: u64, sample_rate: u32) -> Self

return a copy representing the length of time in num_samples