Struct raw::track::Track

pub struct Track { /* fields omitted */ }

A struct to represent a track containing a left and right channel. In future iterations, Track will be able to handle N channels.

Methods

impl Track

pub fn new(name: String, sample_rate: u32) -> Self

Creates a new, empty clip with a given name and sample rate.

Example

let track = Track::new("my_track".to_owned(), 44100);

pub fn sample_rate(&self) -> u32

pub fn name<'a>(&'a self) -> &'a str

pub fn insert_mono<T: Time>(
&mut self,
clip: Arc<Clip>,
position_time: T
) -> bool

Inserts a Clip into self, shifting later clips by the inserted clip's duration.

pub fn insert_stereo(
&mut self,
left_clip: Arc<Clip>,
right_clip: Arc<Clip>,
position: u64
) -> bool

Inserts two clips corresponding to left and right into the track. Clips do not need to be the same length. The clips will start at the same position and the shorter clip is padded with 0s to the duration of the longer clip

pub fn duration(&self) -> u64

Returns the duration of the entire Track. If the mono and stero channels have different durations, the longer one is returned.

pub fn left_channel_as_clip(&self) -> Arc<Clip>

pub fn right_channel_as_clip(&self) -> Arc<Clip>